chore(afe4490): actualizar la libreria a v0.81 y adaptar los consumidores
La dependencia esta fijada a `#master`, asi que "actualizar la version" es
traer el HEAD del repo de la libreria: 0.53.0 (V18) / 0.32.0 (V16, V17) ->
0.81.0 en los tres entornos. Se actualiza solo esa dependencia; las demas
son URLs git sin pin y un `pio pkg update` global habria movido medio
proyecto.

La v0.69 de la libreria migro `ppg_disp` al dominio OT y cambio su tipo de
`int32_t` (cuentas de ADC) a `float` (A/A, magnitud ~1e-5..1e-6). El
firmware seguia compilando sin un solo error porque `float -> int32_t` es
una conversion implicita legal: los cuatro consumidores se rompian en
silencio, truncando cada muestra a 0.

- CommTask: `PPG_DISP_COUNTS_PER_LSB` (16 cuentas/LSB, medido en banco el
  2026-08-13) ya no aplica a estas unidades. Sustituido por
  `PPG_DISP_UNITS_PER_LSB` = 1.6e-8 A/A, DERIVADO de la spec de la libreria
  (DC de OT ~1.4e-5, AC/DC = PI/100, luego pico ~1.4e-6 a PI ~10%), no
  medido — queda un TODO de revalidacion en banco. Se anade guarda de
  `isfinite()`: con un float, un NaN haria indefinido el `lroundf()`, asi
  que colapsa a la linea base en vez de emitir un byte basura. El contrato
  de cable `CTRL,PPG,0..255` no cambia; PROTOCOL.md sigue vigente.
- DriveUpload: `DrivePpgSample::ppg_disp` a float y el CSV pasa de `%d` a
  `%.4e`, el mismo formato que adopto la propia libreria en v0.69.
- PpgSnapshot: buffer y getter a float; el JSON de ThingsBoard crece ~14%
  (~22 KB -> ~25 KB para 400 muestras), sin normalizar aqui.
- SPO2: el log usaba `String(float)` (2 decimales) y habria impreso 0.00
  siempre. Corregidos tambien dos comentarios desfasados (HGAC ya no es
  solo descenso desde v0.81; la referencia a v0.53).

Verificado: build ESP32-S3 SUCCESS en IncuNest_V16/V17/V18; `pio test -e
native` 188/188.

+++ Firmware/motherBoard/src/tasks/PpgSnapshot.cpp
@@ -2,7 +2,9 @@
 
 namespace {
 
-int32_t  s_buf[PPG_SNAPSHOT_SAMPLES];
+// float desde la v0.69 de la libreria: ppg_disp paso al dominio OT (A/A,
+// ~1e-5..1e-6) y como int32_t se truncaba a 0 en todas las muestras.
+float    s_buf[PPG_SNAPSHOT_SAMPLES];
 uint16_t s_count      = 0;     // muestras (ya diezmadas) recogidas
 uint16_t s_decimCount = 0;     // muestras crudas a 500 Hz desde la última guardada
 bool     s_capturing  = false;
@@ -70,7 +72,7 @@ bool ppgSnapshotIsReady() { return s_ready; }
 
 uint16_t ppgSnapshotSampleCount() { return s_count; }
 
-const int32_t *ppgSnapshotSamples() { return s_buf; }
+const float *ppgSnapshotSamples() { return s_buf; }
 
 void ppgSnapshotClear() {
   s_ready = false;
